+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Implementations for the trait `EqStrUntilNul` now allow `?Sized` inputs. This means that
   you can write `some_cstr16.eq_str_until_nul("test")` instead of
   `some_cstr16.eq_str_until_nul(&"test")` now.
+- Added `TryFrom<core::ffi::CStr>` implementation for `CStr8`.
 
 ## uefi-macros - [Unreleased]